refactor(api): extract follows from client.ts
